requirements.txt
streamlit
langchain_google_genai
google-generativeai
langchain_community
faiss-cpu
python-dotenv
typing-extensions



